docintel_phase1_cpu/docintel/requirements.txt
# ==============================================================================
# DocIntel — CPU-only requirements (Dell 5310, i5-10th gen, 16GB RAM)
# Using Google Gemini API instead of Anthropic Claude
# ==============================================================================
# Install order:
#   1. pip install torch --index-url https://download.pytorch.org/whl/cpu
#   2. pip install -r requirements.txt

# ---------- Core schema & config ----------
pydantic>=2.8,<3.0
pydantic-settings>=2.4,<3.0

# ---------- Ingestion ----------
pypdfium2>=4.30
Pillow>=10.4

# ---------- Parsing ----------
pymupdf4llm>=0.0.17
pdfplumber>=0.11
pytesseract>=0.3.13

# ---------- Embedding ----------
sentence-transformers>=3.0
optimum[onnxruntime]>=1.21
numpy>=1.26,<3.0
torch>=2.3,<3.0          # install CPU wheel first (see above)

# ---------- Indexing ----------
qdrant-client>=1.11,<2.0
opensearch-py>=2.6,<3.0

# ---------- Gemini API ----------
google-genai>=1.0          # official Google Gen AI SDK (replaces anthropic)

# ---------- API layer ----------
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0

# ---------- Dev / test ----------
pytest>=8.3
pytest-asyncio>=0.24
ruff>=0.6
